Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float num1;
-    scanf("%f",&num1);
+    double num1;
+    scanf("%lf",&num1);
     printf("You entered: %.4f", num1);
     // printf("%s", welcome());
     return 0;
